chore(explorer, trading): market info shared panels
