+++ libs/market-info/src/components/market-info/index.ts
@@ -3,3 +3,4 @@ export * from './info-market';
 export * from './tooltip-mapping';
 export * from './__generated__/MarketInfo';
 export * from './market-info-data-provider';
+export * from './market-info-panels';
