+++ libs/market-info/src/components/market-info/info-market.tsx
@@ -1,17 +1,5 @@
-import { AssetDetailsTable, useAssetDataProvider } from '@vegaprotocol/assets';
 import { useEnvironment } from '@vegaprotocol/environment';
-import {
-  totalFeesPercentage,
-  calcCandleVolume,
-} from '@vegaprotocol/market-list';
-import {
-  addDecimalsFormatNumber,
-  formatNumber,
-  formatNumberPercentage,
-  removePaginationWrapper,
-  TokenLinks,
-  getMarketExpiryDateFormatted,
-} from '@vegaprotocol/utils';
+import { removePaginationWrapper, TokenLinks } from '@vegaprotocol/utils';
 import { t } from '@vegaprotocol/i18n';
 import { useDataProvider, useYesterday } from '@vegaprotocol/react-helpers';
 import * as Schema from '@vegaprotocol/types';
@@ -22,15 +10,31 @@ import {
   Link as UILink,
   Splash,
 } from '@vegaprotocol/ui-toolkit';
-import BigNumber from 'bignumber.js';
 import { useMemo } from 'react';
 import { generatePath, Link } from 'react-router-dom';
 
-import { MarketInfoTable } from './info-key-value-table';
 import { marketInfoWithDataAndCandlesProvider } from './market-info-data-provider';
 
 import type { MarketInfoWithDataAndCandles } from './market-info-data-provider';
 import { MarketProposalNotification } from '@vegaprotocol/proposals';
+import {
+  CurrentFeesInfoPanel,
+  InstrumentInfoPanel,
+  InsurancePoolInfoPanel,
+  KeyDetailsInfoPanel,
+  LiquidityInfoPanel,
+  LiquidityMonitoringParametersInfoPanel,
+  LiquidityPriceRangeInfoPanel,
+  MarketPriceInfoPanel,
+  MarketVolumeInfoPanel,
+  MetadataInfoPanel,
+  OracleInfoPanel,
+  PriceMonitoringBoundsInfoPanel,
+  RiskFactorsInfoPanel,
+  RiskModelInfoPanel,
+  RiskParametersInfoPanel,
+  SettlementAssetInfoPanel,
+} from './market-info-panels';
 
 export interface InfoProps {
   market: MarketInfoWithDataAndCandles;
@@ -80,15 +84,6 @@ export const MarketInfoContainer = ({
 export const Info = ({ market, onSelect }: InfoProps) => {
   const { VEGA_TOKEN_URL, VEGA_EXPLORER_URL } = useEnvironment();
   const headerClassName = 'uppercase text-lg';
-  const assetSymbol =
-    market?.tradableInstrument.instrument.product?.settlementAsset.symbol || '';
-  const quoteUnit =
-    market?.tradableInstrument.instrument.product?.quoteName || '';
-  const assetId = useMemo(
-    () => market?.tradableInstrument.instrument.product?.settlementAsset.id,
-    [market]
-  );
-  const { data: asset } = useAssetDataProvider(assetId ?? '');
 
   if (!market) return null;
 
@@ -96,272 +91,75 @@ export const Info = ({ market, onSelect }: InfoProps) => {
     market.accountsConnection?.edges
   );
 
-  const last24hourVolume = market.candles && calcCandleVolume(market.candles);
-
   const marketDataPanels = [
     {
       title: t('Current fees'),
-      content: (
-        <>
-          <MarketInfoTable
-            data={{
-              ...market.fees.factors,
-              totalFees: totalFeesPercentage(market.fees.factors),
-            }}
-            asPercentage={true}
-          />
-          <p className="text-xs">
-            {t(
-              'All fees are paid by price takers and are a % of the trade notional value. Fees are not paid during auction uncrossing.'
-            )}
-          </p>
-        </>
-      ),
+      content: <CurrentFeesInfoPanel market={market} />,
     },
     {
       title: t('Market price'),
-      content: (
-        <>
-          <MarketInfoTable
-            data={{
-              markPrice: market.data?.markPrice,
-              bestBidPrice: market.data?.bestBidPrice,
-              bestOfferPrice: market.data?.bestOfferPrice,
-              quoteUnit: market.tradableInstrument.instrument.product.quoteName,
-            }}
-            decimalPlaces={market.decimalPlaces}
-          />
-          <p className="text-xs mt-4">
-            {t(
-              'There is 1 unit of the settlement asset (%s) to every 1 quote unit (%s).',
-              [assetSymbol, quoteUnit]
-            )}
-          </p>
-        </>
-      ),
+      content: <MarketPriceInfoPanel market={market} />,
     },
     {
       title: t('Market volume'),
-      content: (
-        <MarketInfoTable
-          data={{
-            '24hourVolume':
-              last24hourVolume && last24hourVolume !== '0'
-                ? addDecimalsFormatNumber(
-                    last24hourVolume,
-                    market.positionDecimalPlaces
-                  )
-                : '-',
-            openInterest: market.data?.openInterest,
-            bestBidVolume: market.data?.bestBidVolume,
-            bestOfferVolume: market.data?.bestOfferVolume,
-            bestStaticBidVolume: market.data?.bestStaticBidVolume,
-            bestStaticOfferVolume: market.data?.bestStaticOfferVolume,
-          }}
-          decimalPlaces={market.positionDecimalPlaces}
-        />
-      ),
+      content: <MarketVolumeInfoPanel market={market} />,
     },
     ...marketAccounts
       .filter((a) => a.type === Schema.AccountType.ACCOUNT_TYPE_INSURANCE)
       .map((a) => ({
         title: t(`Insurance pool`),
-        content: (
-          <MarketInfoTable
-            data={{
-              balance: a.balance,
-            }}
-            assetSymbol={assetSymbol}
-            decimalPlaces={
-              market.tradableInstrument.instrument.product.settlementAsset
-                .decimals
-            }
-          />
-        ),
+        content: <InsurancePoolInfoPanel market={market} account={a} />,
       })),
   ];
-
-  const keyDetails = {
-    decimalPlaces: market.decimalPlaces,
-    positionDecimalPlaces: market.positionDecimalPlaces,
-    tradingMode: market.tradingMode,
-    state: Schema.MarketStateMapping[market.state],
-  };
-
-  const assetDecimals =
-    market.tradableInstrument.instrument.product.settlementAsset.decimals;
-
-  const liquidityPriceRange = formatNumberPercentage(
-    new BigNumber(market.lpPriceRange).times(100)
-  );
 
   const marketSpecPanels = [
     {
       title: t('Key details'),
-      content: (
-        <MarketInfoTable
-          data={{
-            name: market.tradableInstrument.instrument.name,
-            marketID: market.id,
-            tradingMode:
-              keyDetails.tradingMode &&
-              Schema.MarketTradingModeMapping[keyDetails.tradingMode],
-            marketDecimalPlaces: market.decimalPlaces,
-            positionDecimalPlaces: market.positionDecimalPlaces,
-            settlementAssetDecimalPlaces: assetDecimals,
-          }}
-        />
-      ),
+      content: <KeyDetailsInfoPanel market={market} />,
     },
     {
       title: t('Instrument'),
-      content: (
-        <MarketInfoTable
-          data={{
-            marketName: market.tradableInstrument.instrument.name,
-            code: market.tradableInstrument.instrument.code,
-            productType:
-              market.tradableInstrument.instrument.product.__typename,
-            ...market.tradableInstrument.instrument.product,
-          }}
-        />
-      ),
+      content: <InstrumentInfoPanel market={market} />,
     },
     {
       title: t('Settlement asset'),
-      content: asset ? (
-        <>
-          <AssetDetailsTable
-            asset={asset}
-            inline={true}
-            noBorder={true}
-            dtClassName="text-black dark:text-white text-ui !px-0 !font-normal"
-            ddClassName="text-black dark:text-white text-ui !px-0 !font-normal max-w-full"
-          />
-          <p className="text-xs mt-4">
-            {t(
-              'There is 1 unit of the settlement asset (%s) to every 1 quote unit (%s).',
-              [assetSymbol, quoteUnit]
-            )}
-          </p>
-        </>
-      ) : (
-        <Splash>{t('No data')}</Splash>
-      ),
+      content: <SettlementAssetInfoPanel market={market} />,
     },
     {
       title: t('Metadata'),
-      content: (
-        <MarketInfoTable
-          data={{
-            expiryDate: getMarketExpiryDateFormatted(
-              market.tradableInstrument.instrument.metadata.tags
-            ),
-            ...market.tradableInstrument.instrument.metadata.tags
-              ?.map((tag) => {
-                const [key, value] = tag.split(':');
-                return { [key]: value };
-              })
-              .reduce((acc, curr) => ({ ...acc, ...curr }), {}),
-          }}
-        />
-      ),
+      content: <MetadataInfoPanel market={market} />,
     },
     {
       title: t('Risk model'),
-      content: (
-        <MarketInfoTable
-          data={market.tradableInstrument.riskModel}
-          unformatted={true}
-          omits={[]}
-        />
-      ),
+      content: <RiskModelInfoPanel market={market} />,
     },
     {
       title: t('Risk parameters'),
-      content: (
-        <MarketInfoTable
-          data={market.tradableInstrument.riskModel.params}
-          unformatted={true}
-          omits={[]}
-        />
-      ),
+      content: <RiskParametersInfoPanel market={market} />,
     },
     {
       title: t('Risk factors'),
-      content: (
-        <MarketInfoTable
-          data={market.riskFactors}
-          unformatted={true}
-          omits={['market', '__typename']}
-        />
-      ),
+      content: <RiskFactorsInfoPanel market={market} />,
     },
     ...(market.priceMonitoringSettings?.parameters?.triggers || []).map(
-      (trigger, i) => {
-        const bounds = market.data?.priceMonitoringBounds?.[i];
-        return {
-          title: t(`Price monitoring bounds ${i + 1}`),
-          content: (
-            <div className="text-xs">
-              <div className="grid grid-cols-2 text-xs mb-4">
-                <p className="col-span-1">
-                  {t('%s probability price bounds', [
-                    formatNumberPercentage(
-                      new BigNumber(trigger.probability).times(100)
-                    ),
-                  ])}
-                </p>
-                <p className="col-span-1 text-right">
-                  {t('Within %s seconds', [formatNumber(trigger.horizonSecs)])}
-                </p>
-              </div>
-              <div className="pl-2 pb-0 text-xs border-l-2">
-                {bounds && (
-                  <MarketInfoTable
-                    data={{
-                      highestPrice: bounds.maxValidPrice,
-                      lowestPrice: bounds.minValidPrice,
-                    }}
-                    decimalPlaces={market.decimalPlaces}
-                    assetSymbol={quoteUnit}
-                  />
-                )}
-              </div>
-              <p className="mt-4">
-                {t('Results in %s seconds auction if breached', [
-                  trigger.auctionExtensionSecs.toString(),
-                ])}
-              </p>
-            </div>
-          ),
-        };
-      }
+      (_, triggerIndex) => ({
+        title: t(`Price monitoring bounds ${triggerIndex + 1}`),
+        content: (
+          <PriceMonitoringBoundsInfoPanel
+            market={market}
+            triggerIndex={triggerIndex}
+          />
+        ),
+      })
     ),
     {
       title: t('Liquidity monitoring parameters'),
-      content: (
-        <MarketInfoTable
-          data={{
-            triggeringRatio:
-              market.liquidityMonitoringParameters.triggeringRatio,
-            ...market.liquidityMonitoringParameters.targetStakeParameters,
-          }}
-        />
-      ),
+      content: <LiquidityMonitoringParametersInfoPanel market={market} />,
     },
     {
       title: t('Liquidity'),
       content: (
-        <MarketInfoTable
-          data={{
-            targetStake: market.data && market.data.targetStake,
-            suppliedStake: market.data && market.data?.suppliedStake,
-            marketValueProxy: market.data && market.data.marketValueProxy,
-          }}
-          decimalPlaces={assetDecimals}
-          assetSymbol={assetSymbol}
-        >
+        <LiquidityInfoPanel market={market}>
           <Link
             to={`/liquidity/${market.id}`}
             onClick={() => onSelect(market.id)}
@@ -369,57 +167,17 @@ export const Info = ({ market, onSelect }: InfoProps) => {
           >
             <UILink>{t('View liquidity provision table')}</UILink>
           </Link>
-        </MarketInfoTable>
+        </LiquidityInfoPanel>
       ),
     },
     {
       title: t('Liquidity price range'),
-      content: (
-        <>
-          <p className="text-xs mb-4">
-            {`For liquidity orders to count towards a commitment, they must be
-            within the liquidity monitoring bounds.`}
-          </p>
-          <p className="text-xs mb-4">
-            {`The liquidity price range is a ${liquidityPriceRange} difference from the mid
-            price.`}
-          </p>
-          <div className="pl-2 pb-0 text-xs border-l-2">
-            <MarketInfoTable
-              data={{
-                liquidityPriceRange: `${liquidityPriceRange} of mid price`,
-                lowestPrice:
-                  market.data?.midPrice &&
-                  `${addDecimalsFormatNumber(
-                    new BigNumber(1)
-                      .minus(market.lpPriceRange)
-                      .times(market.data.midPrice)
-                      .toString(),
-                    market.decimalPlaces
-                  )} ${quoteUnit}`,
-                highestPrice:
-                  market.data?.midPrice &&
-                  `${addDecimalsFormatNumber(
-                    new BigNumber(1)
-                      .plus(market.lpPriceRange)
-                      .times(market.data.midPrice)
-                      .toString(),
-                    market.decimalPlaces
-                  )} ${quoteUnit}`,
-              }}
-            ></MarketInfoTable>
-          </div>
-        </>
-      ),
+      content: <LiquidityPriceRangeInfoPanel market={market} />,
     },
     {
       title: t('Oracle'),
       content: (
-        <MarketInfoTable
-          data={
-            market.tradableInstrument.instrument.product.dataSourceSpecBinding
-          }
-        >
+        <OracleInfoPanel market={market}>
           <ExternalLink
             href={`${VEGA_EXPLORER_URL}/oracles#${market.tradableInstrument.instrument.product.dataSourceSpecForSettlementData.id}`}
           >
@@ -430,7 +188,7 @@ export const Info = ({ market, onSelect }: InfoProps) => {
           >
             {t('View termination oracle specification')}
           </ExternalLink>
-        </MarketInfoTable>
+        </OracleInfoPanel>
       ),
     },
   ];
